requirements.txt
asgiref==3.9.1
Django==5.2.4
django-debug-toolbar==6.0.0
djangorestframework==3.16.0
drf-nested-routers==0.94.2
pillow==11.3.0
sqlparse==0.5.3
